+++ script/tool/test/util.dart
@@ -47,16 +47,12 @@ Directory createPackagesDirectory(
 class PlatformDetails {
   const PlatformDetails(
     this.type, {
-    this.variants = const <String>[],
     this.hasNativeCode = true,
     this.hasDartCode = false,
   });
 
   /// The type of support for the platform.
   final PlatformSupport type;
-
-  /// Any 'supportVariants' to list in the pubspec.
-  final List<String> variants;
 
   /// Whether or not the plugin includes native code.
   ///
@@ -291,18 +287,6 @@ String _pluginPlatformSection(
         break;
     }
     entry = lines.join('\n') + '\n';
-  }
-
-  // Add any variants.
-  if (support.variants.isNotEmpty) {
-    entry += '''
-        supportedVariants:
-''';
-    for (final String variant in support.variants) {
-      entry += '''
-          - $variant
-''';
-    }
   }
 
   return entry;